Feature/proposals and data management

See merge request rarimo/rarimo-core!12

+++ x/rarimocore/keeper/msg_server_transfer.go
@@ -81,7 +81,7 @@ func (k msgServer) CreateTransferOperation(goCtx context.Context, msg *types.Msg
 		Details:       details,
 		Signed:        false,
 		Creator:       msg.Creator,
-		Timestamp:     ctx.BlockHeight(),
+		Timestamp:     uint64(ctx.BlockHeight()),
 	}
 
 	k.SetOperation(
